+++ dal/lofar/TBB_File.h
@@ -1,6 +1,8 @@
 #ifndef DAL_TBB_FILE_H
 #define DAL_TBB_FILE_H
 
+#include <stdint.h>
+
 #include <string>
 #include <vector>
 #include <complex>
@@ -96,7 +98,7 @@ public:
   Attribute<unsigned>                   sampleNumber();
 
   Attribute<unsigned>                   samplesPerFrame();
-  Attribute<unsigned>                   dataLength();
+  Attribute<uint64_t>                   dataLength();
   Attribute< std::vector<unsigned> >    flagOffsets();
   Attribute<unsigned>                   nyquistZone();
